Made importBookmarks reusable
