+++ pages/dashboard.tsx
@@ -18,6 +18,7 @@ import Links from "@/components/LinkViews/Links";
 import useLocalSettingsStore from "@/store/localSettings";
 import { useUpdateUser, useUser } from "@/hooks/store/user";
 import SurveyModal from "@/components/ModalContent/SurveyModal";
+import importBookmarks from "@/lib/client/importBookmarks"
 
 export default function Dashboard() {
   const { t } = useTranslation();
@@ -72,70 +73,6 @@ export default function Dashboard() {
       return 2;
     }
   }, []);
-
-  const importBookmarks = async (
-    e: React.ChangeEvent<HTMLInputElement>,
-    format: MigrationFormat
-  ) => {
-    const file: File | null = e.target.files && e.target.files[0];
-
-    if (file) {
-      const reader = new FileReader();
-      //reader.readAsText(file, "UTF-8");
-      reader.readAsArrayBuffer(file);
-      reader.onload = async function (e) {
-        const load = toast.loading("Importing...");
-
-        // const request: string = e.target?.result as string;
-
-        // const body: MigrationRequest = {
-        //   format,
-        //   data: request,
-        // };
-
-        try {
-          const migrationFormatName = MigrationFormat[format];
-          const response = await fetch(`/api/v1/migration/${migrationFormatName}`, {
-            method: "POST",
-            body: e.target?.result as ArrayBuffer,
-            headers: { "Content-Type": "application/octet-stream" },
-          });
-
-          if (!response.ok) {
-            const errorData = await response.json();
-            toast.dismiss(load);
-
-            toast.error(
-              errorData.response ||
-              "Failed to import bookmarks. Please try again."
-            );
-            return;
-          }
-
-          await response.json();
-          toast.dismiss(load);
-          toast.success("Imported the Bookmarks! Reloading the page...");
-
-          setTimeout(() => {
-            location.reload();
-          }, 2000);
-        } catch (error) {
-          console.error("Request failed", error);
-          toast.dismiss(load);
-          toast.error(
-            "An error occurred while importing bookmarks. Please check the logs for more info."
-          );
-        }
-      };
-
-      reader.onerror = function (e) {
-        console.log("Error reading file:", e);
-        toast.error(
-          "Failed to read the file. Please make sure the file is correct and try again."
-        );
-      };
-    }
-  };
 
   const [newLinkModal, setNewLinkModal] = useState(false);
